*Removed the bit selecting contious read mode (line 131 of AD7734.cpp) *Changed the return type of GetConversionData to uint16_t

+++ AD7734.h
@@ -36,7 +36,7 @@ class AD7734 {
     void StartContinousConversion(int adc_channel);
 
     //request channel data from the ADC
-    double GetConversionData(int adc_channel);
+    uint16_t GetConversionData(int adc_channel);
 
     private:
     int _cs, _rdy, _rst;
